--- src/elements/_base.less
@@ -1,9 +1,0 @@
-@import (reference) "vars";
-
-*, *::before, *::after {
-  box-sizing: border-box;
-  // TODO: convert below properties into .inheritTypography() mixin
-  color: inherit;
-  font: inherit;
-  letter-spacing: inherit;
-}
