+++ src/elements/hx-alert/_shadow.scss
@@ -1,4 +1,4 @@
-@import "../_base";
+@import "../base";
 
 #hxAlert {
   display: flex;
