+++ src/elements/hx-div/_shadow.scss
@@ -1,4 +1,4 @@
-@import "../_base";
+@import "../base";
 
 // @supports used to prevent IE from loading the ShadyCSS-rewritten styles.
 // This is so we can write shorter, less-specific CSS selectors in order
