+++ src/elements/hx-drop-fence/_shadow.scss
@@ -1,4 +1,4 @@
-@import "../_base";
+@import "../base";
 
 #hxDropFence {
   align-items: center;
