+++ src/elements/hx-file-icon/_shadow.scss
@@ -1,4 +1,4 @@
-@import "../_base";
+@import "../base";
 
 #hxFileIcon {
   position: relative;
