+++ src/elements/hx-file-input/_shadow.scss
@@ -1,8 +1,4 @@
-@import "../_base";
-
-#hxFileInput {
-  //width: 100%;
-}
+@import "../base";
 
 #hxIcon {
   display: none;
