+++ src/elements/hx-icon/_shadow.scss
