+++ src/elements/hx-pill/_shadow.scss
@@ -1,5 +1,5 @@
-@import "../_base";
-@import (reference) "components/button/index";
+@import "../base";
+@import "components/button/mixins";
 
 :host {
   #hxPill {
@@ -11,9 +11,9 @@
   }
 
   #hxDismiss {
-    #Button.reset();
+    @include button-reset;
     align-items: center;
-    color: var(--hxDismiss-color, @gray-600);
+    color: var(--hxDismiss-color, $gray-600);
     display: inline-flex;
     height: 1.5em;
     justify-content: center;
@@ -25,7 +25,7 @@
     }
 
     &:hover {
-      color: var(--hxDismiss-color--hover, @gray-1000);
+      color: var(--hxDismiss-color--hover, $gray-1000);
     }
   }
 }
