--- src/elements/hx-popover/_shadow.less
@@ -1,10 +1,0 @@
-@import "../_base";
-@import "../_has-arrow";
-
-#hxPopover {
-  background-color: var(--hxBackgroundColor, @gray-0);
-  border-color: var(--hxBorderColor, @gray-300);
-  border-style: solid;
-  border-width: 1px;
-  box-shadow: @layering-shadow;
-}
