+++ src/elements/hx-progress/_shadow.scss
