+++ src/elements/hx-search/_shadow.scss
@@ -1,5 +1,8 @@
-@import (reference) "mixins";
-@import "../_base";
+@import "vars";
+@import "mixins";
+@import "components/form/mixins";
+
+@import "../base";
 
 // TODO: add CSS custom props for theming hooks
 
@@ -20,7 +23,7 @@ input::-ms-clear {
   }
 
   #hxIcon {
-    color: @gray-700;
+    color: $gray-700;
     flex-shrink: 0;
     line-height: 1;
     order: 1;
@@ -29,12 +32,12 @@ input::-ms-clear {
   }
 
   #hxNativeControl {
-    #Mix.placeholder();
-    #Mix.resetFocus();
+    @include hxPlaceholder;
+    @include reset-focus;
 
     background-color: transparent;
     border: none;
-    color: @gray-900;
+    color: $gray-900;
     cursor: inherit;
     flex-grow: 1;
     font-weight: 400;
@@ -45,20 +48,20 @@ input::-ms-clear {
 
     &:focus {
       ~ #hxClear {
-        color: @cyan-700;
+        color: $cyan-700;
       }
 
       ~ #hxCustomControl {
-        border-color: @cyan-700;
-        box-shadow: @focus-glow;
+        border-color: $cyan-700;
+        box-shadow: $focus-glow;
       }
     }
   }
 
   #hxCustomControl {
-    background-color: @gray-0;
+    background-color: $gray-0;
     border-radius: 2px;
-    border: 1px solid @gray-500;
+    border: 1px solid $gray-500;
     height: 100%;
     left: 0;
     position: absolute;
@@ -68,11 +71,11 @@ input::-ms-clear {
   }
 
   #hxClear {
-    #Mix.resetFocus();
+    @include reset-focus;
 
     background-color: transparent;
     border: none;
-    color: @gray-700;
+    color: $gray-700;
     cursor: pointer;
     flex-shrink: 0;
     line-height: 1;
@@ -87,19 +90,19 @@ input::-ms-clear {
       }
 
       ~ * {
-        color: @cyan-700;
+        color: $cyan-700;
       }
 
       ~ #hxCustomControl {
-        border-color: @cyan-700;
-        box-shadow: @focus-glow;
+        border-color: $cyan-700;
+        box-shadow: $focus-glow;
       }
     }//:focus
   }
 }//default
 
 :host([invalid]) {
-  color: @red-900;
+  color: $red-900;
 
   // ----- IDLE -----
 
@@ -109,7 +112,7 @@ input::-ms-clear {
   }
 
   #hxCustomControl {
-    border-color: @red-900;
+    border-color: $red-900;
     border-width: 2px;
   }
 
@@ -123,7 +126,7 @@ input::-ms-clear {
 
   #hxNativeControl:focus {
     ~ #hxClear {
-      color: @red-900;
+      color: $red-900;
     }
   }
 
@@ -132,14 +135,14 @@ input::-ms-clear {
   #hxClear:focus,
   #hxNativeControl:focus {
     ~ #hxCustomControl {
-      box-shadow: @focus-glow-invalid;
-      border-color: @red-900;
+      box-shadow: $focus-glow-invalid;
+      border-color: $red-900;
     }
   }
 }//[invalid]
 
 :host([disabled]) {
-  color: @gray-400;
+  color: $gray-400;
 
   #hxSearch {
     color: inherit;
@@ -155,14 +158,14 @@ input::-ms-clear {
   }
 
   #hxNativeControl {
-    #Mix.disabledPlaceholder();
+    @include hxPlaceholder(disabled);
 
     color: inherit;
   }
 
   #hxCustomControl {
-    background-color: @gray-50;
-    border-color: @gray-300;
+    background-color: $gray-50;
+    border-color: $gray-300;
     border-width: 1px;
   }
 }//[disabled]
