+++ src/elements/hx-tooltip/_shadow.scss
@@ -1,0 +1,13 @@
+@import "vars";
+@import "../base";
+@import "../has-arrow";
+
+#hxTooltip {
+  background-color: var(--hxBackgroundColor, $gray-0);
+  border-color: var(--hxBorderColor, $gray-300);
+  border-radius: 2px;
+  border-style: solid;
+  border-width: 1px;
+  box-shadow: $layering-shadow;
+  padding: 0.75rem;
+}
